backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
transformers==4.53.0
torch==2.9.1
pydantic==2.11.7
openai==1.52.0
google-generativeai==0.8.3
pillow==11.1.0
python-multipart==0.0.12
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
redis==5.2.0
markdown==3.7
bleach==6.2.0
requests==2.32.4
numpy==2.2.6
pandas==2.2.6
langchain==0.3.8
langchain-community==0.3.8
langchain-openai==0.2.8
sentence-transformers==3.3.1
chromadb==0.5.23
pypdf2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5
httpx==0.27.0
aiohttp==3.9.0
